[SA] [DdpmSwUpdate]
1.MiniInstaller rename
2.Software add check json content
3.Add return

+++ DDPM.SA/Plugins/DDPM.SA.Plugins.User.DeviceManager/DeviceManagerPlugin.cs
@@ -3721,31 +3721,35 @@ namespace DDPM.SA.Plugins.User.DeviceManager
                 }
             }
         }
-        private void DeleteMiniInstallerFolder()
+        private void DeleteDdpmSwUpdaterFolder()
         {
-            writelog("[DeleteMiniInstallerFolder], start.");
+            writelog("[DeleteDdpmSwUpdaterFolder], start.");
             string registryKey = @"SOFTWARE\Dell Display and Peripheral Manager";
-            object o = ReadRegistryData(RegistryHive.LocalMachine, registryKey, "MiniInstaller").Result;
-            writelog($"[DeleteMiniInstallerFolder], o={o}.");
+            object o = ReadRegistryData(RegistryHive.LocalMachine, registryKey, "DdpmSwUpdater").Result;
+            writelog($"[DeleteDdpmSwUpdaterFolder], o={o}.");
             if (o != null && o is string && !string.IsNullOrEmpty(o.ToString()))
             {
-                writelog($"[DeleteMiniInstallerFolder], o_String={o.ToString()}.");
+                writelog($"[DeleteDdpmSwUpdaterFolder], o_String={o.ToString()}.");
                 DDPMFileSecurity DDPMFileSecurity = new DDPMFileSecurity();
                 string AppDataPath = DDPMFileSecurity.GetActiveUserLocalAppDataPath();
                 if (!string.IsNullOrEmpty(AppDataPath))
                 {
                     string path = AppDataPath + "\\Dell\\Dell Display and Peripheral Manager" + "\\" + o.ToString();
-                    if (Directory.Exists(path))
+                    if (!string.IsNullOrEmpty(path) && Directory.Exists(path))
                     {
-                        writelog($"[DeleteMiniInstallerFolder], Exists.");
+                        writelog($"[DeleteDdpmSwUpdaterFolder], Exists.");
                         Directory.Delete(path, true);
-                        writelog($"[DeleteMiniInstallerFolder], Delete.");
+                        writelog($"[DeleteDdpmSwUpdaterFolder], Delete.");
                     }
-                    WriteRegistryData(RegistryHive.LocalMachine, registryKey, "MiniInstaller", "");
-                    writelog($"[DeleteMiniInstallerFolder], WriteRegistryData.");
+                    WriteRegistryData(RegistryHive.LocalMachine, registryKey, "DdpmSwUpdater", "");
+                    writelog($"[DeleteDdpmSwUpdaterFolder], WriteRegistryData.");
+                }
+                else
+                {
+                    writelog("[DeleteDdpmSwUpdaterFolder], AppDataPath get null.");
                 }
             }
-            writelog("[DeleteMiniInstallerFolder], done.");
+            writelog("[DeleteDdpmSwUpdaterFolder], done.");
         }
         #endregion
 
@@ -4728,7 +4732,7 @@ namespace DDPM.SA.Plugins.User.DeviceManager
             DDMMigration();
             ReloadHotkeyConfigData();
             ToNKVM_initHotKeys();
-            DeleteMiniInstallerFolder();
+            DeleteDdpmSwUpdaterFolder();
             GetSkipCA().Wait();
             //hook keyboard
             //if (_HotkeyPlugin != null)
